move websocket interface to ocenet, cleanup
